+++ fsic/core/models.py
@@ -116,7 +116,12 @@ class BaseModel(SolverMixin, ModelInterface):
         return cls(index, *args, **{k: v.values for k, v in data.items()}, **kwargs)
 
     def reindex(
-        self, span: Sequence[Hashable], *, fill_value: Any = None, **fill_values: Any
+        self,
+        span: Sequence[Hashable],
+        *,
+        fill_value: Any = None,
+        strict: Optional[bool] = None,
+        **fill_values: Any,
     ) -> 'BaseModel':
         """Return a copy of the current object, adjusted to match `span`. Values in overlapping periods between the old and new objects are preserved (copied over).
 
@@ -124,6 +129,11 @@ class BaseModel(SolverMixin, ModelInterface):
         ----------
         span : iterable
             Sequence of periods defining the span of the object to be returned
+        strict : bool
+            If `True`, raise a `KeyError` if `fill_values` refers to variables
+            not defined in the current object. Ignore if `False`.
+            If `None`, use the current value of the object's `strict`
+            attribute.
         fill_value :
             Default fill value for new periods
         **fill_values :
@@ -141,7 +151,9 @@ class BaseModel(SolverMixin, ModelInterface):
         fill_values['status'] = fill_values.get('status', SolutionStatus.UNSOLVED.value)
         fill_values['iterations'] = fill_values.get('iterations', -1)
 
-        return super().reindex(span, fill_value=fill_value, **fill_values)
+        return super().reindex(
+            span, fill_value=fill_value, strict=strict, **fill_values
+        )
 
     def to_dataframe(
         self,