+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 - Added new keyword argument `include_internal` to control
   `model_to_dataframe()` handling of internal model variables with names
   prefixed with an underscore.
+- Added new `PandasIndexFeaturesMixin` extension for more sophisticated `span`
+  handling.
 - Added new `TracerMixin` extension for `BaseModel`-derived classes, to track
   iteration-by-iteration variable results.